requirements.txt
aniso8601==9.0.1
blinker==1.7.0
bokeh==2.4.3
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.0.3
colorama==0.4.4
contourpy==1.0.5
cycler==0.11.0
datatest==0.11.1
DateTime==4.3
docopt==0.6.2
et-xmlfile==1.1.0
Flask==2.0.2
Flask-RESTful==0.3.9
fonttools==4.37.4
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.4.4
MarkupSafe==2.0.1
matplotlib==3.6.1
more-itertools==9.0.0
mpld3==0.5.8
notifications-python-client==6.3.0
numpy==1.26.1
opencv-python==4.6.0.66
openpyxl==3.0.9
packaging==21.3
pandas==2.1.2
pathlib==1.0.1
patsy==0.5.3
Pillow==9.2.0
PyJWT==2.4.0
pyparsing==3.0.9
pypng==0.20220715.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzbar==0.1.9
qrcode==7.3.1
Quandl==3.7.0
requests==2.28.1
scipy==1.11.3
six==1.16.0
statsmodels==0.14.0
tornado==6.2
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.11
Werkzeug==2.0.2
xyzservices==2023.10.1
zope.interface==5.4.0
